Move EnumLike type class to the macros project
