+++ macros/src/main/scala/io/sigs/seals/macros/EnumLike.scala
@@ -15,7 +15,7 @@
  */
 
 package io.sigs.seals
-package core
+package macros
 
 import scala.reflect.macros.blackbox.Context
